chore: rearrange packages into monorepo
