+++ .eslintrc
@@ -1,0 +1,19 @@
+{
+  "extends": "eslint:recommended",
+  "parser": "babel-eslint",
+  "parserOptions": {
+    "sourceType": "module"
+  },
+  "env": {
+    "node": true,
+    "es6": true
+  },
+  "overrides": [
+    {
+      "files": ["**/test/**"],
+      "env": {
+        "jest": true
+      }
+    }
+  ]
+}
