--- packages/icss-utils/.travis.yml
@@ -1,36 +1,0 @@
-sudo: false
-
-git:
-  depth: 10
-
-branches:
-  only:
-    - master
-    - next
-
-language: node_js
-
-cache:
-  yarn: true
-  directories:
-    - node_modules
-
-matrix:
-  include:
-    - node_js: "11"
-      env: JOB_PART=pretest
-    - node_js: "6"
-      env: JOB_PART=test
-    - node_js: "8"
-      env: JOB_PART=test
-    - node_js: "10"
-      env: JOB_PART=test
-    - node_js: "11"
-      env: JOB_PART=test
-
-before_script:
-  - node --version
-  - yarn --version
-
-script:
-  - yarn $JOB_PART
