+++ packages/icss-utils/package.json
@@ -10,47 +10,10 @@
     "lib"
   ],
   "scripts": {
-    "lint": "eslint . --ignore-path .gitignore",
-    "build": "babel --out-dir lib src",
-    "pretest": "npm run lint",
-    "test": "npm run test:only",
-    "test:only": "jest",
-    "prepublish": "yarn test && yarn run build"
-  },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": 6
-          }
-        }
-      ]
-    ]
-  },
-  "eslintConfig": {
-    "parser": "babel-eslint",
-    "parserOptions": {
-      "sourceType": "module"
-    },
-    "env": {
-      "es6": true,
-      "jest": true
-    },
-    "extends": "eslint:recommended"
-  },
-  "lint-staged": {
-    "*.js": [
-      "prettier --write",
-      "eslint",
-      "git add"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+    "build": "babel --root-mode upward --out-dir lib src",
+    "test": "jest --coverage",
+    "test:watch": "jest --watch",
+    "prepublishOnly": "yarn test && yarn run build"
   },
   "repository": {
     "type": "git",
@@ -73,14 +36,6 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.1.0",
-    "@babel/core": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.1.0",
-    "eslint": "^5.14.1",
-    "husky": "^1.3.1",
-    "jest": "^24.1.0",
-    "lint-staged": "^8.1.4",
-    "prettier": "^1.16.4"
+    "@babel/core": "^7.1.0"
   }
 }
