--- packages/postcss-modules-local-by-default/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-npm-debug.log
-coverage
