--- packages/postcss-modules-scope/.eslintrc
@@ -1,11 +1,0 @@
-{
-	"parserOptions": {
-		"ecmaVersion": 2018
-	},
-	"env": {
-		"node": true
-	},
-	"rules": {
-		"quotes": [2, "single"]
-	}
-}
