--- packages/postcss-modules-scope/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-coverage
-lib
