--- packages/postcss-modules-scope/.travis.yml
@@ -1,12 +1,0 @@
-language: node_js
-node_js:
-  - "6"
-  - "8"
-  - "10"
-  - "12"
-script: npm run travis
-
-after_success:
-  - cat ./coverage/lcov.info | node_modules/.bin/coveralls --verbose
-  - cat ./coverage/coverage.json | node_modules/codecov.io/bin/codecov.io.js
-  - rm -rf ./coverage
